Implement product mutations and API helpers
