+++ backend/src/PosBackend/Application/Requests/UpdateProductRequest.cs
@@ -1,0 +1,5 @@
+namespace PosBackend.Application.Requests;
+
+public sealed class UpdateProductRequest : ProductMutationRequestBase
+{
+}
