+++ backend/src/PosBackend/Features/Products/ProductsController.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
@@ -9,6 +10,7 @@ using PosBackend.Application.Requests;
 using PosBackend.Application.Responses;
 using PosBackend.Application.Services;
 using PosBackend.Infrastructure.Data;
+using PosBackend.Domain.Entities;
 
 namespace PosBackend.Features.Products;
 
@@ -20,14 +22,18 @@ public class ProductsController : ControllerBase
     private readonly ApplicationDbContext _db;
     private readonly MlClient _mlClient;
     private readonly ScanWatchdog _watchdog;
-    private readonly ProductService _productService;
+    private readonly CurrencyService _currencyService;
 
-    public ProductsController(ApplicationDbContext db, MlClient mlClient, ScanWatchdog watchdog, ProductService productService)
+    public ProductsController(
+        ApplicationDbContext db,
+        MlClient mlClient,
+        ScanWatchdog watchdog,
+        CurrencyService currencyService)
     {
         _db = db;
         _mlClient = mlClient;
         _watchdog = watchdog;
-        _productService = productService;
+        _currencyService = currencyService;
     }
 
     [HttpGet]
@@ -39,7 +45,7 @@ public class ProductsController : ControllerBase
             .OrderBy(p => p.Name)
             .ToListAsync(cancellationToken);
 
-        var responses = products.Select(_productService.ToResponse).ToList();
+        var responses = products.Select(ToResponse).ToList();
         return Ok(responses);
     }
 
@@ -56,60 +62,96 @@ public class ProductsController : ControllerBase
             return NotFound();
         }
 
-        var response = _productService.ToResponse(product);
+        var response = ToResponse(product);
         return Ok(response);
     }
 
     [HttpPost]
     public async Task<ActionResult<ProductResponse>> CreateProduct([FromBody] CreateProductRequest request, CancellationToken cancellationToken)
     {
-        var result = await _productService.CreateAsync(request, cancellationToken);
-        if (!result.Succeeded)
+        var (errors, category) = await ValidateAsync(request, null, cancellationToken);
+        if (errors.Count > 0)
         {
-            if (result.IsNotFound)
-            {
-                return NotFound();
-            }
-
-            foreach (var error in result.Errors)
-            {
-                foreach (var message in error.Value)
-                {
-                    ModelState.AddModelError(error.Key, message);
-                }
-            }
-
-            return ValidationProblem(ModelState);
+            return CreateValidationProblem(errors);
         }
 
-        var response = _productService.ToResponse(result.Product!);
+        var priceResult = await TryResolvePricesAsync(request, cancellationToken);
+        if (!priceResult.Succeeded)
+        {
+            return CreateValidationProblem(priceResult.Errors);
+        }
+
+        var product = new Product
+        {
+            Name = request.Name!,
+            Sku = request.Sku!,
+            Barcode = request.Barcode!,
+            Description = NormalizeOptional(request.Description),
+            CategoryId = request.CategoryId,
+            PriceUsd = priceResult.PriceUsd,
+            PriceLbp = priceResult.PriceLbp,
+            IsActive = true
+        };
+
+        _db.Products.Add(product);
+        await _db.SaveChangesAsync(cancellationToken);
+
+        product.Category = category;
+        var response = ToResponse(product);
         return CreatedAtAction(nameof(GetProductById), new { id = response.Id }, response);
     }
 
     [HttpPut("{id:guid}")]
-    public async Task<ActionResult<ProductResponse>> UpdateProduct(Guid id, [FromBody] CreateProductRequest request, CancellationToken cancellationToken)
+    public async Task<ActionResult<ProductResponse>> UpdateProduct(Guid id, [FromBody] UpdateProductRequest request, CancellationToken cancellationToken)
     {
-        var result = await _productService.UpdateAsync(id, request, cancellationToken);
-        if (result.IsNotFound)
+        var product = await _db.Products.Include(p => p.Category)
+            .FirstOrDefaultAsync(p => p.Id == id, cancellationToken);
+        if (product is null)
         {
             return NotFound();
         }
 
-        if (!result.Succeeded)
+        var (errors, category) = await ValidateAsync(request, id, cancellationToken);
+        if (errors.Count > 0)
         {
-            foreach (var error in result.Errors)
-            {
-                foreach (var message in error.Value)
-                {
-                    ModelState.AddModelError(error.Key, message);
-                }
-            }
-
-            return ValidationProblem(ModelState);
+            return CreateValidationProblem(errors);
         }
 
-        var response = _productService.ToResponse(result.Product!);
+        var priceResult = await TryResolvePricesAsync(request, cancellationToken);
+        if (!priceResult.Succeeded)
+        {
+            return CreateValidationProblem(priceResult.Errors);
+        }
+
+        product.Name = request.Name!;
+        product.Sku = request.Sku!;
+        product.Barcode = request.Barcode!;
+        product.Description = NormalizeOptional(request.Description);
+        product.CategoryId = request.CategoryId;
+        product.PriceUsd = priceResult.PriceUsd;
+        product.PriceLbp = priceResult.PriceLbp;
+        product.Category = category;
+        product.UpdatedAt = DateTime.UtcNow;
+
+        await _db.SaveChangesAsync(cancellationToken);
+
+        var response = ToResponse(product);
         return Ok(response);
+    }
+
+    [HttpDelete("{id:guid}")]
+    public async Task<IActionResult> DeleteProduct(Guid id, CancellationToken cancellationToken)
+    {
+        var product = await _db.Products.FindAsync(new object?[] { id }, cancellationToken);
+        if (product is null)
+        {
+            return NotFound();
+        }
+
+        _db.Products.Remove(product);
+        await _db.SaveChangesAsync(cancellationToken);
+
+        return NoContent();
     }
 
     [HttpGet("search")]
@@ -128,7 +170,7 @@ public class ProductsController : ControllerBase
             .AsNoTracking()
             .ToListAsync(cancellationToken);
 
-        var results = products.Select(_productService.ToResponse).ToList();
+        var results = products.Select(ToResponse).ToList();
 
         return Ok(results);
     }
@@ -163,5 +205,212 @@ public class ProductsController : ControllerBase
             IsFlagged = flagged,
             FlagReason = reason
         };
+    }
+
+    private ProductResponse ToResponse(Product product)
+    {
+        return new ProductResponse
+        {
+            Id = product.Id,
+            Sku = product.Sku,
+            Name = product.Name,
+            Barcode = product.Barcode,
+            PriceUsd = product.PriceUsd,
+            PriceLbp = product.PriceLbp,
+            Description = product.Description,
+            Category = product.Category?.Name ?? string.Empty
+        };
+    }
+
+    private async Task<(Dictionary<string, string[]>, Category?)> ValidateAsync(
+        ProductMutationRequestBase request,
+        Guid? existingProductId,
+        CancellationToken cancellationToken)
+    {
+        var errors = new Dictionary<string, List<string>>(StringComparer.Ordinal);
+
+        if (string.IsNullOrWhiteSpace(request.Name))
+        {
+            AddError(errors, nameof(request.Name), "Name is required.");
+        }
+        else
+        {
+            request.Name = request.Name.Trim();
+        }
+
+        if (string.IsNullOrWhiteSpace(request.Sku))
+        {
+            AddError(errors, nameof(request.Sku), "SKU is required.");
+        }
+        else
+        {
+            request.Sku = request.Sku.Trim();
+        }
+
+        if (string.IsNullOrWhiteSpace(request.Barcode))
+        {
+            AddError(errors, nameof(request.Barcode), "Barcode is required.");
+        }
+        else
+        {
+            request.Barcode = request.Barcode.Trim();
+        }
+
+        if (request.Price is null)
+        {
+            AddError(errors, nameof(request.Price), "Price is required.");
+        }
+        else if (request.Price.Value < 0)
+        {
+            AddError(errors, nameof(request.Price), "Price cannot be negative.");
+        }
+
+        var currency = NormalizeCurrency(request.Currency);
+        if (currency is null)
+        {
+            AddError(errors, nameof(request.Currency), "Currency must be either USD or LBP.");
+        }
+        else
+        {
+            request.Currency = currency;
+        }
+
+        if (request.CategoryId == Guid.Empty)
+        {
+            AddError(errors, nameof(request.CategoryId), "CategoryId is required.");
+        }
+
+        if (!errors.ContainsKey(nameof(request.Sku)))
+        {
+            var skuExists = await _db.Products
+                .AnyAsync(p => p.Sku == request.Sku && (!existingProductId.HasValue || p.Id != existingProductId.Value), cancellationToken);
+            if (skuExists)
+            {
+                AddError(errors, nameof(request.Sku), "SKU must be unique.");
+            }
+        }
+
+        if (!errors.ContainsKey(nameof(request.Barcode)))
+        {
+            var barcodeExists = await _db.Products
+                .AnyAsync(p => p.Barcode == request.Barcode && (!existingProductId.HasValue || p.Id != existingProductId.Value), cancellationToken);
+            if (barcodeExists)
+            {
+                AddError(errors, nameof(request.Barcode), "Barcode must be unique.");
+            }
+        }
+
+        Category? category = null;
+        if (!errors.ContainsKey(nameof(request.CategoryId)))
+        {
+            category = await _db.Categories
+                .FirstOrDefaultAsync(c => c.Id == request.CategoryId, cancellationToken);
+            if (category is null)
+            {
+                AddError(errors, nameof(request.CategoryId), "Category not found.");
+            }
+        }
+
+        var materialized = errors.ToDictionary(
+            pair => pair.Key,
+            pair => pair.Value.ToArray(),
+            StringComparer.Ordinal);
+
+        return (materialized, category);
+    }
+
+    private async Task<PriceComputationResult> TryResolvePricesAsync(ProductMutationRequestBase request, CancellationToken cancellationToken)
+    {
+        if (request.Price is not decimal priceValue)
+        {
+            return PriceComputationResult.Failure(new Dictionary<string, string[]>
+            {
+                [nameof(request.Price)] = new[] { "Price is required." }
+            });
+        }
+
+        if (priceValue < 0)
+        {
+            return PriceComputationResult.Failure(new Dictionary<string, string[]>
+            {
+                [nameof(request.Price)] = new[] { "Price cannot be negative." }
+            });
+        }
+
+        try
+        {
+            var rate = await _currencyService.GetCurrentRateAsync(cancellationToken);
+            var currency = request.Currency ?? "USD";
+
+            decimal priceUsd;
+            decimal priceLbp;
+
+            if (currency == "LBP")
+            {
+                priceLbp = _currencyService.RoundLbp(priceValue);
+                priceUsd = _currencyService.ConvertLbpToUsd(priceLbp, rate.Rate);
+            }
+            else
+            {
+                priceUsd = _currencyService.RoundUsd(priceValue);
+                priceLbp = _currencyService.ConvertUsdToLbp(priceUsd, rate.Rate);
+            }
+
+            return PriceComputationResult.Success(priceUsd, priceLbp);
+        }
+        catch (InvalidOperationException)
+        {
+            return PriceComputationResult.Failure(new Dictionary<string, string[]>
+            {
+                [nameof(request.Currency)] = new[] { "Exchange rate is not configured." }
+            });
+        }
+    }
+
+    private ActionResult CreateValidationProblem(IDictionary<string, string[]> errors)
+    {
+        ModelState.Clear();
+        foreach (var error in errors)
+        {
+            foreach (var message in error.Value)
+            {
+                ModelState.AddModelError(error.Key, message);
+            }
+        }
+
+        return ValidationProblem(ModelState);
+    }
+
+    private static void AddError(Dictionary<string, List<string>> errors, string key, string message)
+    {
+        if (!errors.TryGetValue(key, out var list))
+        {
+            list = new List<string>();
+            errors[key] = list;
+        }
+
+        list.Add(message);
+    }
+
+    private static string? NormalizeOptional(string? value) => string.IsNullOrWhiteSpace(value) ? null : value.Trim();
+
+    private static string? NormalizeCurrency(string? currency)
+    {
+        if (string.IsNullOrWhiteSpace(currency))
+        {
+            return "USD";
+        }
+
+        var trimmed = currency.Trim().ToUpperInvariant();
+        return trimmed is "USD" or "LBP" ? trimmed : null;
+    }
+
+    private sealed record PriceComputationResult(bool Succeeded, decimal PriceUsd, decimal PriceLbp, IDictionary<string, string[]> Errors)
+    {
+        public static PriceComputationResult Success(decimal priceUsd, decimal priceLbp) =>
+            new(true, priceUsd, priceLbp, new Dictionary<string, string[]>(StringComparer.Ordinal));
+
+        public static PriceComputationResult Failure(IDictionary<string, string[]> errors) =>
+            new(false, 0m, 0m, errors);
     }
 }
