+++ backend/src/PosBackend/Program.cs
@@ -21,7 +21,6 @@ builder.Services.AddScoped<CartPricingService>();
 builder.Services.AddScoped<ReceiptRenderer>();
 builder.Services.AddScoped<AuditLogger>();
 builder.Services.AddScoped<CurrencyService>();
-builder.Services.AddScoped<ProductService>();
 builder.Services.AddHttpClient<MlClient>();
 
 var connectionString = builder.Configuration.GetConnectionString("DefaultConnection")
